[IK-30] refactor:  병합 후 코드 정리
