+++ src/test/java/com/kdt/instakyuram/member/service/MemberServiceIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.kdt.instakyuram.member;
+package com.kdt.instakyuram.member.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
